finished refactoring vet visits
