+++ src/components/medicals/vetVisits/VetVisitsList.js
@@ -1,46 +1,26 @@
 import React, { useEffect, useState } from "react";
 import MedicalRepository from "../../../repositories/MedicalRepository";
 import ContactsRepository from "../../../repositories/ContactsRepository";
-import { AddVetVisitDialog } from "./AddVetVisitDialog"
+import { AddVetVisitDialog } from "./AddVetVisitDialog";
 import Container from "@mui/material/Container";
 import Box from "@mui/material/Box";
-import Card from "@mui/material/Card";
-import CardContent from "@mui/material/CardContent";
-import Button from "@mui/material/Button";
-import CardHeader from "@mui/material/CardHeader";
+import Typography from "@mui/material/Typography";
 import { VetVisit } from "./VetVisit";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import FormControl from "@mui/material/FormControl";
-import InputLabel from "@mui/material/InputLabel";
-import Select from "@mui/material/Select";
-import MenuItem from "@mui/material/MenuItem";
-import TextField from "@mui/material/TextField";
 
-export const VetVisitsList = ({ pet, dashboardView }) => {
+export const VetVisitsList = ({ pet }) => {
     const [myPetVetVisits, setMyPetVetVisits] = useState([]);
     const [vets, setVets] = useState([]);
     const [expanded, setExpanded] = useState(false);
 
     const syncVetVisits = () => {
         MedicalRepository.getAllVetVisitsByPet(pet.id).then((data) => {
-            if (dashboardView) {
-                data = data.filter((vetVisit) => vetVisit.starred);
-                setMyPetVetVisits(data);
-            } else {
-                setMyPetVetVisits(data);
-            }
+            setMyPetVetVisits(data);
         });
     };
 
     useEffect(() => {
-        ContactsRepository.getVetContacts().then((data) => setVets(data));
         syncVetVisits();
-        return () => {
-            setMyPetVetVisits([]);
-        };
+        ContactsRepository.getVetContacts().then((data) => setVets(data));
     }, []);
 
     const handleChange = (panel) => (event, isExpanded) => {
@@ -50,23 +30,24 @@ export const VetVisitsList = ({ pet, dashboardView }) => {
     return (
         <Container maxWidth="lg">
             <Box sx={{ textAlign: "center" }}>
-                <AddVetVisitDialog pet={pet} vets={vets} syncVetVisits={syncVetVisits}/>
-                <Card sx={{ minWidth: 200 }}>
-                    <CardContent>
-                        <CardHeader title={`${pet ? pet.name + "'s ": ""} Vet Visits`}/>
-                        
-                        {myPetVetVisits.map((vetVisit) => (
-                            <VetVisit
-                                key={vetVisit.id}
-                                vetVisit={vetVisit}
-                                syncVetVisits={syncVetVisits}
-                                handleChange={handleChange}
-                                expanded={expanded}
-                                vets={vets}
-                            />
-                        ))}
-                    </CardContent>
-                </Card>
+                <AddVetVisitDialog
+                    pet={pet}
+                    vets={vets}
+                    syncVetVisits={syncVetVisits}
+                />
+                <Typography variant="h5" gutterBottom align="center">
+                    Vet Visits
+                </Typography>
+                {myPetVetVisits.map((vetVisit) => (
+                    <VetVisit
+                        key={vetVisit.id}
+                        vetVisit={vetVisit}
+                        syncVetVisits={syncVetVisits}
+                        handleChange={handleChange}
+                        expanded={expanded}
+                        vets={vets}
+                    />
+                ))}
             </Box>
         </Container>
     );
